+++ scheduling-app/src/App.js
@@ -1,7 +1,6 @@
 import logo from "./logo.svg";
 import "./App.css";
-import CoursePage from "./components/course-page";
-
+import CoursePage from "./components/pages/course-page";
 function App() {
   return (
     <div>
