+++ scheduling-app/src/components/common/form.jsx
@@ -1,0 +1,10 @@
+import React, { Component } from "react";
+
+class Form extends Component {
+  state = {};
+  render() {
+    return <div>Form</div>;
+  }
+}
+
+export default Form;
