+++ scheduling-app/src/components/page-components/course-table.jsx
